myt-magic-binder/requirements.txt
flask
flask_sqlalchemy
werkzeug
psycopg2
requests